# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Google Ads Service",
    "description": "Wraps the Google Ads API (v23) for MCC agency management. Handles OAuth, account linking, campaign CRUD, and performance reporting.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "components": {
    "securitySchemes": {
      "serviceKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Service-to-service API key"
      }
    },
    "schemas": {
...
      "CampaignsResponse": {
        "type": "object",
        "properties": {
          "campaigns": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "ENABLED",
                    "PAUSED",
                    "REMOVED"
                  ]
                },
                "advertisingChannelType": {
                  "type": "string"
                },
                "biddingStrategy": {
                  "type": "string"
                },
                "budgetAmountMicros": {
                  "type": "string"
                },
                "startDate": {
                  "type": "string"
                },
                "endDate": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "name",
                "status",
                "advertisingChannelType"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "campaigns"
        ],
        "additionalProperties": false
      },
...
      "CampaignDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ENABLED",
              "PAUSED",
              "REMOVED"
            ]
          },
          "advertisingChannelType": {
            "type": "string"
          },
          "biddingStrategy": {
            "type": "string"
          },
          "budgetAmountMicros": {
            "type": "string"
          },
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          },
          "resourceName": {
            "type": "string"
          },
          "urlCustomParameters": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "key": {
                  "type": "string"
                },
                "value": {
                  "type": "string"
                }
              },
              "required": [
                "key",
                "value"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "id",
          "name",
          "status",
          "advertisingChannelType",
          "resourceName"
        ],
        "additionalProperties": false
      },
      "PerformanceResponse": {
        "type": "object",
        "properties": {
          "campaignId": {
            "type": "string"
          },
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          },
          "metrics": {
            "type": "object",
            "properties": {
              "impressions": {
                "type": "number"
              },
              "clicks": {
                "type": "number"
              },
              "conversions": {
                "type": "number"
              },
              "costMicros": {
                "type": "string"
              },
              "cpa": {
                "type": "number",
                "nullable": true
              },
              "roas": {
                "type": "number",
                "nullable": true
              },
              "ctr": {
                "type": "number"
              },
              "averageCpc": {
                "type": "number",
                "nullable": true
              }
            },
            "required": [
              "impressions",
              "clicks",
              "conversions",
              "costMicros",
              "cpa",
              "roas",
              "ctr",
              "averageCpc"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "campaignId",
          "startDate",
          "endDate",
          "metrics"
        ],
        "additionalProperties": false
      },
...
      "ConversionsResponse": {
        "type": "object",
        "properties": {
          "conversionActions": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "category": {
                  "type": "string"
                },
                "status": {
                  "type": "string"
                },
                "type": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "name",
                "category",
                "status",
                "type"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "conversionActions"
        ],
        "additionalProperties": false
      },
...
      "CreateCampaignBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "advertisingChannelType": {
            "type": "string",
            "minLength": 1
          },
          "status": {
            "type": "string",
            "enum": [
              "ENABLED",
              "PAUSED",
              "REMOVED"
            ],
            "default": "PAUSED"
          },
          "budgetAmountMicros": {
            "type": "string",
            "minLength": 1
          },
          "biddingStrategy": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
          },
          "endDate": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
          }
        },
        "required": [
          "name",
          "advertisingChannelType",
          "budgetAmountMicros"
        ],
        "additionalProperties": false
      },
      "CreateCampaignResponse": {
        "type": "object",
        "properties": {
          "campaign": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "status": {
                "type": "string",
                "enum": [
                  "ENABLED",
                  "PAUSED",
                  "REMOVED"
                ]
              },
              "advertisingChannelType": {
                "type": "string"
              },
              "biddingStrategy": {
                "type": "string"
              },
              "budgetAmountMicros": {
                "type": "string"
              },
              "startDate": {
                "type": "string"
              },
              "endDate": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "name",
              "status",
              "advertisingChannelType"
            ],
            "additionalProperties": false
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "campaign",
          "message"
        ],
        "additionalProperties": false
      },
      "UpdateCampaignBody": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ENABLED",
              "PAUSED",
              "REMOVED"
            ]
          },
          "budgetAmountMicros": {
            "type": "string"
          },
          "biddingStrategy": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "UpdateCampaignResponse": {
        "type": "object",
        "properties": {
          "campaign": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "status": {
                "type": "string",
                "enum": [
                  "ENABLED",
                  "PAUSED",
                  "REMOVED"
                ]
              },
              "advertisingChannelType": {
                "type": "string"
              },
              "biddingStrategy": {
                "type": "string"
              },
              "budgetAmountMicros": {
                "type": "string"
              },
              "startDate": {
                "type": "string"
              },
              "endDate": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "name",
              "status",
              "advertisingChannelType"
            ],
            "additionalProperties": false
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "campaign",
          "message"
        ],
        "additionalProperties": false
      },
      "DuplicateCampaignBody": {
        "type": "object",
        "properties": {
          "newName": {
            "type": "string",
            "minLength": 1
          }
        },
        "additionalProperties": false
      },
      "DuplicateCampaignResponse": {
        "type": "object",
        "properties": {
          "campaign": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "status": {
                "type": "string",
                "enum": [
                  "ENABLED",
                  "PAUSED",
                  "REMOVED"
                ]
              },
              "advertisingChannelType": {
                "type": "string"
              },
              "biddingStrategy": {
                "type": "string"
              },
              "budgetAmountMicros": {
                "type": "string"
              },
              "startDate": {
                "type": "string"
              },
              "endDate": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "name",
              "status",
              "advertisingChannelType"
            ],
            "additionalProperties": false
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "campaign",
          "message"
        ],
        "additionalProperties": false
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      }
    },
    "parameters": {
      "OrgId": {
        "name": "x-org-id",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Internal org UUID from client-service"
      },
      "UserId": {
        "name": "x-user-id",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Internal user UUID from client-service"
      },
      "RunId": {
        "name": "x-run-id",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Caller's run ID from runs-service (used as parentRunId when creating a child run)"
      },
      "FeatureSlug": {
        "name": "x-feature-slug",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "Feature slug for tracking which feature triggered the request"
      }
    }
  },
  "paths": {
...
    "/accounts/{accountId}/campaigns": {
      "get": {
        "summary": "List campaigns for a Google Ads account",
        "parameters": [
          {
            "$ref": "#/components/parameters/OrgId"
          },
          {
            "$ref": "#/components/parameters/UserId"
          },
          {
            "$ref": "#/components/parameters/RunId"
          },
          {
            "$ref": "#/components/parameters/FeatureSlug"
          },
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "ENABLED",
                "PAUSED",
                "REMOVED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of campaigns",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new campaign",
        "parameters": [
          {
            "$ref": "#/components/parameters/OrgId"
          },
          {
            "$ref": "#/components/parameters/UserId"
          },
          {
            "$ref": "#/components/parameters/RunId"
          },
          {
            "$ref": "#/components/parameters/FeatureSlug"
          },
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCampaignBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Campaign created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCampaignResponse"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/{accountId}/campaigns/{campaignId}": {
      "get": {
        "summary": "Get campaign details",
        "parameters": [
          {
            "$ref": "#/components/parameters/OrgId"
          },
          {
            "$ref": "#/components/parameters/UserId"
          },
          {
            "$ref": "#/components/parameters/RunId"
          },
          {
            "$ref": "#/components/parameters/FeatureSlug"
          },
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "campaignId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Campaign details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignDetail"
                }
              }
            }
          },
          "404": {
            "description": "Campaign not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update campaign (budget, bids, status)",
        "parameters": [
          {
            "$ref": "#/components/parameters/OrgId"
          },
          {
            "$ref": "#/components/parameters/UserId"
          },
          {
            "$ref": "#/components/parameters/RunId"
          },
          {
            "$ref": "#/components/parameters/FeatureSlug"
          },
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "campaignId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCampaignBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Campaign updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateCampaignResponse"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/{accountId}/campaigns/{campaignId}/performance": {
      "get": {
        "summary": "Get campaign performance metrics (impressions, clicks, conversions, CPA, ROAS)",
        "parameters": [
          {
            "$ref": "#/components/parameters/OrgId"
          },
          {
            "$ref": "#/components/parameters/UserId"
          },
          {
            "$ref": "#/components/parameters/RunId"
          },
          {
            "$ref": "#/components/parameters/FeatureSlug"
          },
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "campaignId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Performance metrics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerformanceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/{accountId}/campaigns/{campaignId}/duplicate": {
      "post": {
        "summary": "Duplicate a campaign for A/B testing",
        "parameters": [
          {
            "$ref": "#/components/parameters/OrgId"
          },
          {
            "$ref": "#/components/parameters/UserId"
          },
          {
            "$ref": "#/components/parameters/RunId"
          },
          {
            "$ref": "#/components/parameters/FeatureSlug"
          },
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "campaignId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DuplicateCampaignBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Campaign duplicated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DuplicateCampaignResponse"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/{accountId}/conversions": {
      "get": {
        "summary": "List conversion actions for an account",
        "parameters": [
          {
            "$ref": "#/components/parameters/OrgId"
          },
          {
            "$ref": "#/components/parameters/UserId"
          },
          {
            "$ref": "#/components/parameters/RunId"
          },
          {
            "$ref": "#/components/parameters/FeatureSlug"
          },
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of conversion actions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversionsResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}